Move middleware .ts file to Pylance/node subdir
